% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oaxaca.R
\name{make_decomp_dofile}
\alias{make_decomp_dofile}
\title{Create a \code{.do} file to run a command and save estimates}
\usage{
make_decomp_dofile(command, do_path, dta_file, est_file)
}
\arguments{
\item{command}{String of Stata estimation command.}

\item{do_path}{Path to \code{.do} file to save.}

\item{dta_file}{Data file for Stata to use in same directory as
\code{.do} file.}

\item{est_file}{Filename (without extension) Stata to save estimate results in same
directory as \code{.do} file.}
}
\value{
Nothing, called only for its side effect of writing the
  \code{.do} file.
}
\description{
Internal function to write a \code{.do} file that will run a
command and export a file suitable for reading with
\code{read_stata_estimates}.
}
