% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oaxaca.R
\name{read_stata_estimates}
\alias{read_stata_estimates}
\title{Read estimates from Stata's Excel export}
\usage{
read_stata_estimates(path, type)
}
\arguments{
\item{path}{The path to the Excel file.}

\item{type}{Either "twofold" or "threefold".}
}
\value{
A data frame with the same rows and columns as would be
  produced in the \code{varlevel} element of the output of
  \code{OaxacaBlinderDecomp} (though the rows might be in a
  different order).
}
\description{
This internal function reads the estimates produced by running
Jann's \code{oaxaca} command in Stata.  The estimates should be
exported to a file with extension \code{xlsx} or \code{xls}
using Stata using the \code{etable} command, with the option
\code{cstat(_r_b)} to ensure that standard errors are not
included.
}
