% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oaxaca.R
\name{OaxacaBlinderDecomp}
\alias{OaxacaBlinderDecomp}
\title{OaxacaBlinderDecomp}
\usage{
OaxacaBlinderDecomp(formula, data, type = "twofold")
}
\arguments{
\item{formula:}{formula object specifying the model. The formula should be 
specified as: dep_var ~ x_var1 + x_var2 + ... x_varK | grouping_variable
factor variable are supported by default and can be inserted in the formula
without encoding manually}

\item{data:}{a data.frame}

\item{type:}{the decomposition type: `twofold` or `threefold`}
}
\description{
OaxacaBlinderDecomp
}
